style: put BaseTestClass in __init__.py for use by all tests
